Updated unit tests accordingly
